backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.10.0
python-multipart==0.0.6
# modal==1.0.5  # Removed - using Cloud Run instead
# torch>=2.6.0  # Removed - not needed for API server
# transformers>=4.36.2  # Removed - not needed for API server
# diffusers>=0.25.0  # Removed - not needed for API server
# huggingface-hub>=0.19.4  # Removed - not needed for API server
numpy>=1.24.3
scipy>=1.11.4
biopython>=1.81
# rdkit-pypi>=2023.9.1  # Commented out for Docker build compatibility
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0
toml>=0.10.2
sqlalchemy>=2.0.23
alembic>=1.13.1
psycopg2-binary>=2.9.9
redis>=5.0.1
celery>=5.3.4
google-cloud-storage>=2.10.0
google-cloud-firestore>=2.11.0
google-cloud-tasks>=2.13.0
google-cloud-pubsub>=2.31.0
google-cloud-compute>=1.12.0
google-auth>=2.17.0
google-cloud-monitoring>=2.15.0
google-cloud-functions>=1.13.0
prometheus_client>=0.17.0
PyJWT>=2.8.0
aiohttp>=3.8.0
passlib>=1.7.4
python-jose[cryptography]>=3.3.0
httpx>=0.24.0
asyncpg>=0.27.0
aiocache>=0.12.0
croniter>=1.3.0
tenacity>=8.2.0
structlog>=23.1.0
sentry-sdk>=1.40.0 